Fix public API docs, publicapigen
