+++ runtime/beta/auth/auth.go
@@ -14,17 +14,16 @@ import (
 // UID is a unique identifier representing a user (a user id).
 type UID = model.UID
 
+//publicapigen:drop
 type Manager struct {
 	rt *reqtrack.RequestTracker
 }
 
+//publicapigen:drop
 func NewManager(rt *reqtrack.RequestTracker) *Manager {
 	return &Manager{rt}
 }
 
-// UserID reports the uid of the user making the request.
-// The second result is true if there is a user and false
-// if the request was made without authentication details.
 func (mgr *Manager) UserID() (UID, bool) {
 	if curr := mgr.rt.Current(); curr.Req != nil {
 		return curr.Req.UID, curr.Req.UID != ""
@@ -32,15 +31,6 @@ func (mgr *Manager) UserID() (UID, bool) {
 	return "", false
 }
 
-// Data returns the structured auth data for the request.
-// It returns nil if the request was made without authentication details,
-// and the API endpoint does not require them.
-//
-// Expected usage is to immediately cast it to the registered auth data type:
-//
-//   usr, ok := auth.Data().(*user.Data)
-//   if !ok { /* ... */ }
-//
 func (mgr *Manager) Data() interface{} {
 	if curr := mgr.rt.Current(); curr.Req != nil {
 		return curr.Req.AuthData
