+++ runtime/pubsub/package.go
@@ -1,5 +1,5 @@
 // Package pubsub provides Encore applications with the ability to create Topics and multilpe Subscriptions on those
 // topics in a cloud-agnostic manner.
 //
-// For more information see http://encore.dev/docs/develop/pubsub
+// For more information see https://encore.dev/docs/develop/pubsub
 package pubsub
